Moved to CMake. Modules feature updates.
view: Moved ScrollView from Controls 1 to Controls 2.
workspace.Project: Moved responsibility of Runnable.Trigger to document.
editqmljs: Added ComponentSource.
base: Added QmlPromise.
base: Added QmlFunctionObjectCall.
base.Act: Added support for QmlPromise.
base: Added StreamCollector.
base: Added TextConcat.
base: Added StreamOperations.
base: ByteBuffer dependency fix.
base: FileSystem include fix.
base: Added StreamOperations.
base: Added ValueFlowGraph.
base: Removed GroupCollector, GroupCollectorEntry, EachValueApply.
base: Added ValueFlowGraphPalette.
base: Added ValueFlowGraph.
editor: Updated BindingChannel views.
fs: Added DirectoryToStream.
live: Removed StaticContainer and StaticLoader.
elements.Compiler: Added file change check before recompiling.
Removed QStringRef.
Replaced QLinkedList with std::list.
lcvcore: Moved alphaMerge, roi and overlap to MatOp functions.
lcvvideo: Remoevd StaticContainer usage.
workspace: Added builder module.

+++ lib/lvelements/src/languageinfo.cpp
@@ -480,7 +480,7 @@ void TypeInfo::fromMLNode(const MLNode &node)
 
     MLNode prop = object.at("properties");
     MLNode::ArrayType prop_array = prop.asArray();
-    for (int i = 0; i < prop_array.size(); ++i)
+    for (int i = 0; i < static_cast<int>(prop_array.size()); ++i)
     {
         MLNode property = prop_array[i];
         PropertyInfo pi("");
